// src/services/rest-api/customer-service/model/fatca-w9-input.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.267
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface FatcaW9Input
 */
export interface FatcaW9Input {
    /**
     * 
     * @type {string}
     * @memberof FatcaW9Input
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW9Input
     */
    'businessName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW9Input
     */
    'appropriate'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW9Input
     */
    'address1'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW9Input
     */
    'address2'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW9Input
     */
    'address3'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW9Input
     */
    'zipCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW9Input
     */
    'accountList'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW9Input
     */
    'ssn'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW9Input
     */
    'ein'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW9Input
     */
    'exemptionPayeeCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW9Input
     */
    'exemptionReportingCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW9Input
     */
    'refId'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW9Input
     */
    'refType'?: string | null;
}